+++ src/app.js
@@ -1,6 +1,6 @@
 import Fastify from 'fastify';
-import { getDB, saveDB } from './utils/db/index.js';
-import { v4 } from 'uuid';
+import { general } from './services/general/index.js';
+import { createBlog } from './services/blogs/create-blog.js';
 
 const prefix = '/api';
 
@@ -8,35 +8,9 @@ export async function build () {
   // initialize fastify
   const fastify = Fastify({ logger: true });
 
-  fastify.get(prefix, async (request, reply) => {
-    return { success: true };
-  });
+  fastify.get(prefix, general);
 
-  // create blog
-  fastify.post(`${prefix}/blog`, async (request, reply) => {
-    const { body } = request;
-    const { title, description, isDone = false } = body;
-    const db = await getDB();
-
-    const id = v4();
-
-    const blog = {
-      title,
-      description,
-      isDone,
-      createdDate: new Date().getTime(),
-      updatedDate: new Date().getTime()
-    };
-
-    db.blogs[id] = blog;
-
-    await saveDB(db);
-
-    return {
-      id,
-      ...blog
-    };
-  });
+  fastify.post(`${prefix}/blog`, createBlog);
 
   return fastify;
 }
